pyproject.toml
[project]
name = "omg-cli"
version = "0.1.0"
description = "A CLI for OMG (Observability, Monitoring, and Governance) tools powered by LLMs."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.86.0",
    "anyio>=4.11.0",
    "arclet-letoderea>=0.20.7",
    "dotenv>=0.9.9",
    "fastmcp>=3.2.0",
    "inline-snapshot>=0.31.0",
    "loguru>=0.7.3",
    "markdown-it-py[plugins]>=4.0.0",
    "mistletoe>=1.5.0",
    "mistune>=3.1.4",
    "openai>=2.7.1",
    "pathspec>=0.12.1",
    "pydantic>=2.12.3",
    "pygtrie>=2.5.0",
    "pytodotxt>=3.1.0",
    "pyyaml>=6.0.3",
    "textual-dev>=1.8.0",
    "tomli-w>=1.2.0",
    "typer>=0.20.0",
]

[dependency-groups]
dev = ["pytest>=8.4.2", "pytest-asyncio>=1.3.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 120

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "TID",   # flake8-tidy-imports
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
    "T201",   # print-statement
]


[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["src", "tests/*"]
extra-standard-library = ["typing_extensions"]


[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
